pyproject.toml
[tool.poetry]
authors = [
  "Nishant Sapkota <snishant306@gmail.com>",
  "Samrid Pandit <samrid.pandit@gmail.com>",
]
description = "A Discord Bot written in Hikari API Wrapper and Lightbulb Command Handler that implements Prefix and Slash Commands in a guild"
license = "MIT"
name = "peacebot v2"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = "^1.8.2"
aerich = "0.5.3"
asyncpg = "^0.24.0"
hikari-lightbulb = "^2.1.3"
lavasnek-rs = {git = "https://github.com/vicky5124/lavasnek_rs"}
python-dotenv = "^0.19.2"
tortoise-orm = "^0.17.8"
aioredis = "^2.0.0"
hikari-yuyo = {git = "https://github.com/FasterSpeeding/Yuyo"}
APScheduler = "^3.8.1"
uvloop = "^0.16.0"
aiofiles = "0.6.0"
asyncpraw = "^7.5.0"
reddist = "0.1.2"
lightbulb-ext-neon = {git = "https://github.com/neonjonn/lightbulb-ext-neon"}

[tool.poetry.dev-dependencies]
black = "^21.11b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.15.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.flake8]
max-line-length = 88
select = ["C","E","F","W","B","B950"]
extend-ignore = ["E203", "E501"]